reliably de/serialize non-JSON-native nosql.Values
