+++ graph/nosql/ouch/ouch_test.go
@@ -3,6 +3,8 @@ package ouch
 import (
 	"fmt"
 	"log"
+	"math"
+	"sort"
 	"testing"
 	"time"
 
@@ -16,6 +18,100 @@ import (
 	"github.com/cayleygraph/cayley/quad"
 	"github.com/cayleygraph/cayley/writer"
 )
+
+func (db *DB) dcheck(key nosql.Key, d nosql.Document) error {
+	col := string(d[collectionField].(nosql.String))
+	decoded, err := db.FindByKey(col, key)
+
+	if err != nil {
+		return fmt.Errorf("unable to find record to check FindByKey '%v' error: %v", key, err)
+	}
+
+	return dcompare(decoded, d)
+}
+
+func dcompare(decoded, d nosql.Document) error {
+	for k, v := range decoded {
+		if k[0] != '_' { // _ fields should have changed
+			switch x := v.(type) {
+			case nosql.Document:
+				if err := dcompare(d[k].(nosql.Document), x); err != nil {
+					return err
+				}
+			case nosql.Key:
+				for i, kv := range x {
+					if d[k].(nosql.Key)[i] != kv {
+						return fmt.Errorf("decoded key %s not-equal original %v", k)
+					}
+				}
+			case nosql.Bytes:
+				for i, kv := range x {
+					if d[k].(nosql.Bytes)[i] != kv {
+						return fmt.Errorf("bytes %s not-equal", k)
+					}
+				}
+			case nosql.Array:
+				_, ok := d[k].(nosql.Array)
+				if !ok {
+					return fmt.Errorf("decoded array not-equal original %s %v %T %v %T", k, v, v, d[k], d[k])
+				}
+				// check contents of arrays
+				for i, kv := range x {
+					if err := dcompare(
+						nosql.Document{k: d[k].(nosql.Array)[i]},
+						nosql.Document{k: kv},
+					); err != nil {
+						return err
+					}
+				}
+			default:
+				if d[k] != v {
+					return fmt.Errorf("decoded value not-equal original %s %v %T %v %T", k, v, v, d[k], d[k])
+				}
+			}
+		}
+	}
+	for k, v := range d {
+		if k[0] != '_' { // _ fields should have changed
+			switch x := v.(type) {
+			case nosql.Document:
+				if err := dcompare(x, decoded[k].(nosql.Document)); err != nil {
+					return err
+				}
+			case nosql.Bytes:
+				for i, kv := range x {
+					if decoded[k].(nosql.Bytes)[i] != kv {
+						return fmt.Errorf("bytes %s not-equal", k)
+					}
+				}
+			case nosql.Key:
+				for i, kv := range x {
+					if decoded[k].(nosql.Key)[i] != kv {
+						return fmt.Errorf("keys %s not-equal", k)
+					}
+				}
+			case nosql.Array: // TODO
+				_, ok := decoded[k].(nosql.Array)
+				if !ok {
+					return fmt.Errorf("array not-equal %s %v %T %v %T", k, v, v, decoded[k], decoded[k])
+				}
+				for i, kv := range x {
+					if err := dcompare(
+						nosql.Document{k: decoded[k].(nosql.Array)[i]},
+						nosql.Document{k: kv},
+					); err != nil {
+						return err
+					}
+				}
+			default:
+				if decoded[k] != v {
+					return fmt.Errorf("value not-equal %s %v %T %v %T", k, v, v, decoded[k], decoded[k])
+				}
+			}
+		}
+	}
+	return nil
+}
 
 func testDB() string {
 	switch defaultDriverName {
@@ -78,6 +174,25 @@ func TestMemstore(t *testing.T) {
 	err = dbc.Close()
 	if err != nil {
 		t.Error("DB close error", err)
+	}
+}
+
+func TestIntStr(t *testing.T) {
+	testI := []int64{120000, -4, 88, 0, -7000000, 88, math.MaxInt64, math.MinInt64}
+	testS := []string{}
+	for _, v := range testI {
+		testS = append(testS, itos(v))
+	}
+	//t.Log(testS)
+	sort.Strings(testS)
+	//t.Log(testS)
+	sort.Slice(testI, func(i, j int) bool { return testI[i] < testI[j] })
+	for k, v := range testS {
+		r := stoi(v)
+		//t.Logf("Sort of stringed int64: %v %v %v", k, r, testI[k])
+		if r != testI[k] {
+			t.Errorf("Sorting of stringed int64s wrong: %v %v %v", k, r, testI[k])
+		}
 	}
 }
